+++ src/store/pages/modules/pageRace.js
@@ -53,6 +53,27 @@ export default {
 		},
 	},
 	actions: {
+    showRaceSetings({ commit, dispatch, state, rootState }, {name, key}){
+      console.log("showRace", name, key)
+      let common = rootState.MY.MY.ethnos.name === "common";
+			let ethnos_common = (name === "ethnos" && common);
+
+      let color_length = rootState.MY.MY.race.settings.color[key]?.length === 0;
+			let color_common = (name === `${key}_color` && color_length);
+			if (ethnos_common || color_common) {
+				return null;
+			} else if (state.shown[name] === false) {
+        dispatch("closeSetings", {page: "main_page"});
+				dispatch("closeSetings", {page: "race_page"});
+        commit("OPEN_SETING", {page: "race_page", name: name});
+				// commit("OPEN_RACE_SETING", name);
+				commit("CLOSE_HOME");
+
+			} else {
+				dispatch("goHome");
+			}
+		},
+
 		closeRaceSetings({ commit, state }) {
 			const obj = state.shown;
 			for (const [key, value] of Object.entries(obj)) {
